split kubernetes package in common/kubelet/apiserver to reduce dsd size
