+++ pkg/util/kubernetes/kubelet/types_kubelet.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2017 Datadog, Inc.
 
-package kubernetes
+package kubelet
 
 // Pod contains fields for unmarshalling a Pod
 type Pod struct {
